refactor(core): extract coding scope resolution
